trekk ut testcontainer oppsett fra integrasjonTestOppsett
