+++ src/test/kotlin/no/nav/helse/flex/testconfig/IntegrasjonTestOppsett.kt
@@ -5,7 +5,6 @@ import no.nav.helse.flex.arbeidsforhold.ArbeidsforholdRepository
 import no.nav.helse.flex.narmesteleder.NarmesteLederRepository
 import no.nav.helse.flex.sykmelding.SykmeldingRepository
 import no.nav.helse.flex.tsmsykmeldingstatus.SykmeldingStatusBufferRepository
-import no.nav.helse.flex.utils.logger
 import no.nav.security.token.support.spring.test.EnableMockOAuth2Server
 import org.apache.kafka.clients.producer.Producer
 import org.junit.jupiter.api.AfterAll
@@ -19,15 +18,6 @@ import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.kafka.config.KafkaListenerEndpointRegistry
 import org.springframework.kafka.test.utils.ContainerTestUtils
 import org.springframework.test.web.servlet.MockMvc
-import org.testcontainers.containers.GenericContainer
-import org.testcontainers.containers.PostgreSQLContainer
-import org.testcontainers.kafka.KafkaContainer
-import org.testcontainers.utility.DockerImageName
-
-// TODO: Bytt fra latest til en konkret versjon når det er tilgjengelig i dockerhub
-private class ValkeyContainer : GenericContainer<ValkeyContainer>("bitnamisecure/valkey:latest")
-
-private class PostgreSQLContainer14 : PostgreSQLContainer<PostgreSQLContainer14>("postgres:14-alpine")
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @AutoConfigureObservability
@@ -41,7 +31,7 @@ private class PostgreSQLContainer14 : PostgreSQLContainer<PostgreSQLContainer14>
     ],
 )
 @AutoConfigureMockMvc(print = MockMvcPrint.NONE, printOnlyOnFailure = false)
-abstract class IntegrasjonTestOppsett {
+abstract class IntegrasjonTestOppsett : TestcontainersTestOppsett() {
     @Autowired
     lateinit var mockMvc: MockMvc
 
@@ -62,51 +52,6 @@ abstract class IntegrasjonTestOppsett {
 
     @Autowired
     lateinit var kafkaListenerEndpointRegistry: KafkaListenerEndpointRegistry
-
-    companion object {
-        private val logger = logger()
-
-        init {
-            KafkaContainer(DockerImageName.parse("apache/kafka-native:3.8.1")).apply {
-                start()
-                System.setProperty("KAFKA_BROKERS", bootstrapServers)
-
-                logger.info(
-                    "Started kafka testcontainer: bootstrapServers=${this.bootstrapServers}, " +
-                        "image=${this.dockerImageName}, containerId=${this.containerId.take(12)}",
-                )
-            }
-
-            PostgreSQLContainer14().apply {
-                withCommand("postgres", "-c", "wal_level=logical")
-                start()
-                System.setProperty("spring.datasource.url", "$jdbcUrl&reWriteBatchedInserts=true")
-                System.setProperty("spring.datasource.username", username)
-                System.setProperty("spring.datasource.password", password)
-
-                logger.info(
-                    "Started Postgres testcontainer: jdbcUrl=${this.jdbcUrl}, " +
-                        "image=${this.dockerImageName}, containerId=${this.containerId.take(12)}",
-                )
-            }
-
-            ValkeyContainer().apply {
-                withEnv("ALLOW_EMPTY_PASSWORD", "yes")
-                withExposedPorts(6379)
-                start()
-
-                System.setProperty("VALKEY_HOST_SESSIONS", host)
-                System.setProperty("VALKEY_PORT_SESSIONS", firstMappedPort.toString())
-                System.setProperty("VALKEY_USERNAME_SESSIONS", "default")
-                System.setProperty("VALKEY_PASSWORD_SESSIONS", "")
-
-                logger.info(
-                    "Started Velkey testcontainer: host=${this.host}, port=${this.firstMappedPort}, " +
-                        "image=${this.dockerImageName}, containerId=${this.containerId.take(12)}",
-                )
-            }
-        }
-    }
 
     @BeforeAll
     fun beforeAllFelles() {
